third_party/blink/web_tests/platform/generic/http/tests/misc/webtiming-one-redirect-expected.txt

This test checks that Web Timing reports redirectCount correctly and has reasonable values for redirectStart and redirectCount.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timing.redirectStart is >= timing.navigationStart
PASS timing.redirectEnd is >= timing.redirectStart
PASS navigation.redirectCount is 1
PASS successfullyParsed is true

TEST COMPLETE
